Moved i686 cpu stuff into own folder. Added implementaion for idt and isr's. isr handler still missing.

+++ build.bat
@@ -1,3 +1,3 @@
 @echo off
 if "%~1"=="1" docker build buildenv -t sodium-buildenv
-docker run --rm -it -v %cd%:/root/env sodium-buildenv
+if "%errorlevel%" == "0" docker run --rm -it -v %cd%:/root/env sodium-buildenv
